+++ pyvo/utils/__init__.py
@@ -1,2 +1,3 @@
 from .compat import *
 from .prototype import prototype_feature, activate_features
+from .download import http_download, aws_download